Adding tests for new timezone support comparison function
